Extract `DeliveryStatusView` into a separate file
